fix: serialization type sizes

+++ lapcounter/lapcounter.c
@@ -284,7 +284,7 @@ int lc_save(const lc_counter_t* counter, const char* path) {
   fwrite(&counter->start_inclination_vector.y, sizeof(double), 1, f);
   fwrite(&counter->start_line_vector.x, sizeof(double), 1, f);
   fwrite(&counter->start_line_vector.y, sizeof(double), 1, f);
-  fwrite(&counter->sector_count, sizeof(double), 1, f);
+  fwrite(&counter->sector_count, sizeof(int), 1, f);
   for(int i = 0; i < counter->sector_count; i++) {
     fwrite(&counter->points[i].x, sizeof(double), 1, f);
     fwrite(&counter->points[i].y, sizeof(double), 1, f);
@@ -309,7 +309,7 @@ int lc_load(lc_counter_t* counter, const char* path){
   fread(&counter->start_inclination_vector.y, sizeof(double), 1, f);
   fread(&counter->start_line_vector.x, sizeof(double), 1, f);
   fread(&counter->start_line_vector.y, sizeof(double), 1, f);
-  fread(&counter->sector_count, sizeof(double), 1, f);
+  fread(&counter->sector_count, sizeof(int), 1, f);
   _lc_init_sectors(counter, counter->sector_count);
   for(int i = 0; i < counter->sector_count; i++) {
     fread(&counter->points[i].x, sizeof(double), 1, f);
